split constraint files
