+++ test/constraint/encoding.test.ts
@@ -9,7 +9,9 @@ import {Type} from 'vega-lite/build/src/type';
 import {Property} from '../../src/property';
 import {PropIndex} from '../../src/propindex';
 import {DEFAULT_QUERY_CONFIG} from '../../src/config';
-import {EncodingConstraintModel, FIELD_CONSTRAINTS, FIELD_CONSTRAINT_INDEX, VALUE_CONSTRAINT_INDEX} from '../../src/constraint/encoding';
+import {EncodingConstraintModel} from '../../src/constraint/base';
+import {FIELD_CONSTRAINTS, FIELD_CONSTRAINT_INDEX} from '../../src/constraint/field';
+import {VALUE_CONSTRAINT_INDEX} from '../../src/constraint/value';
 import {EncodingQuery, ScaleQuery, FieldQuery, ValueQuery} from '../../src/query/encoding';
 import {SHORT_WILDCARD, Wildcard} from '../../src/wildcard';
 import {duplicate, extend} from '../../src/util';
